Move rhings around, parseReroll
