+++ src/digestTotals/cap/capDigester.ts
