+++ src/digestTotals/cap/singleCapDigester.ts
