+++ src/digestTotals/drop/dropDigester.ts
