+++ src/digestTotals/explode/explodeDigester.ts
