+++ src/digestTotals/unique/uniqueDigester.ts
